techDebt(buRefactorTraining): Move csvQuote into bulkUploadUtils
